+++ CHANGELOG.md
@@ -13,6 +13,8 @@ cider-nrepl's info middleware for jump-to-definition.
 * New interactive command `cider-insert-defun-in-repl`.
 * New interactive command `cider-insert-ns-form-in-repl`.
 * New inspector inspired by SLIME's inspector
+* STDERR ouput is now font-locked with `cider-repl-err-output-face` to make it
+visually distinctive from `cider-repl-output-face` (used for STDOUT output).
 
 ### Changes